requirements.txt
openai~=1.44.1
python-dotenv~=1.0.1
pydantic~=2.9.0